feat: move boards resourses to ts
