+++ package.json
@@ -47,6 +47,7 @@
     "cross-env": "7.0.3",
     "dotenv": "10.0.0",
     "fastify": "3.24.1",
+    "fastify-plugin": "^3.0.0",
     "fastify-swagger": "4.12.6",
     "http-status-codes": "2.1.4",
     "uuid": "8.3.2",
@@ -54,6 +55,7 @@
   },
   "devDependencies": {
     "@types/node": "17.0.0",
+    "@types/uuid": "^8.3.3",
     "@typescript-eslint/eslint-plugin": "5.7.0",
     "@typescript-eslint/parser": "5.7.0",
     "chai": "4.3.4",
