--- src/resources/boards/board.memory.repository.js
@@ -1,15 +1,0 @@
-const DB = require('../../db/inMemoryDb');
-
-const TABLE_NAME = 'Boards';
-
-const getAll = async () => DB.getAllEntities(TABLE_NAME);
-
-const getById = async (id) => DB.getEntity(TABLE_NAME, id);
-
-const add = async (entity) => DB.addEntity(TABLE_NAME, entity);
-
-const remove = async (id) => DB.deleteEntity(TABLE_NAME, id);
-
-const update = async (id, entity) => DB.updateEntity(TABLE_NAME, id, entity);
-
-module.exports = { add, getAll, getById, remove, update };
