--- src/resources/boards/column.model.js
@@ -1,20 +1,0 @@
-const uuid = require('uuid');
-
-class Column {
-  constructor({
-    id = uuid.v4(),
-    title = 'My Board',
-    order = 0
-  } = {}) {
-    this.id = id;
-    this.title = title;
-    this.order = order;
-  }
-
-  static toResponse(column) {
-    const { id, title, order } = column;
-    return { id, title, order };
-  }
-}
-
-module.exports = Column;
